working on editor v2
