+++ boot.py
@@ -8,9 +8,14 @@ if __name__ == "__main__":
         editor.save(True)
 
     if "-e" in sys.argv or "--editor" in sys.argv:
-        from src import editor
-        G = editor.set_up()
-        editor.run(G)
+        if "-v2" in sys.argv:
+            from src import editor_v2
+            G = editor_v2.set_up()
+            editor_v2.run(G)
+        else:
+            from src import editor
+            G = editor.set_up()
+            editor.run(G)
     elif "-t" in sys.argv:
         from src import tas
         G = tas.set_up()
