+++ src/lib/ACTORS.py
@@ -2679,6 +2679,13 @@ ACTORS = {'BW_WOce_cm0': {'DIM': (592, 464),
                'scripts': 'CAcell64',
                'sprites': 'CAcell64',
                'tangible': False},
+ 'CAgrid160': {'DIM': (144, 304),
+               'POS': (704, 256),
+               'name': 'CAgrid160',
+               'offsetkey': 'CAgrid16',
+               'scripts': 'CAgrid16',
+               'sprites': 'CAgrid16',
+               'tangible': False},
  'CAgrid320': {'DIM': (800, 640),
                'POS': (656, -288),
                'name': 'CAgrid320',
@@ -2707,6 +2714,20 @@ ACTORS = {'BW_WOce_cm0': {'DIM': (592, 464),
              'scripts': 'CArule',
              'sprites': 'CArule',
              'tangible': False},
+ 'CArule1': {'DIM': (16, 16),
+             'POS': (672, 176),
+             'name': 'CArule1',
+             'offsetkey': 'CArule',
+             'scripts': 'CArule',
+             'sprites': 'CArule',
+             'tangible': False},
+ 'CArule_0': {'DIM': (16, 16),
+              'POS': (720, 176),
+              'name': 'CArule_0',
+              'offsetkey': 'CArule_',
+              'scripts': 'CArule_',
+              'sprites': 'CArule_',
+              'tangible': False},
  'WOcastle0_plains21': {'DIM': (352, 416),
                         'POS': (-27584, -160),
                         'name': 'WOcastle0_plains21',
