+++ src/lib/WORLDS.py
@@ -1142,14 +1142,14 @@ WORLDS = {'BW_castleentrence': {'actors': ['BW_grave10',
              'name': 'castle0',
              'x_lock': None,
              'y_lock': None},
- 'city0': {'actors': ['platformskyscrape8',
+ 'city0': {'actors': ['grocerystore0',
+                      'platformskyscrape8',
                       'platform43',
                       'platformroad0',
                       'WOcity0_yard20',
                       'platformroad1',
                       'platformskyscrape0',
                       'platformskyscrape1',
-                      'grocerystore0',
                       'spike20',
                       'spike22',
                       'inviswall0',
@@ -4263,7 +4263,10 @@ WORLDS = {'BW_castleentrence': {'actors': ['BW_grave10',
                      'gooseball5',
                      'cheatermodedisplay0',
                      'puppetredpants0',
-                     'playsong_house0'],
+                     'playsong_house0',
+                     'CAgrid160',
+                     'CArule1',
+                     'CArule_0'],
           'background': 'wooden',
           'name': 'root',
           'x_lock': None,
